+++ forge/src/main/java/com/iamkaf/amber/networking/forge/ForgeClientNetworking.java
@@ -13,6 +13,12 @@ import net.minecraftforge.fml.loading.FMLLoader;
  */
 public class ForgeClientNetworking {
 
+    static net.minecraft.network.Connection connection() {
+        var client = net.minecraft.client.Minecraft.getInstance();
+        var listener = client.getConnection();
+        return listener == null || client.player == null ? null : listener.getConnection();
+    }
+
     static Player getLocalPlayer() {
         return net.minecraft.client.Minecraft.getInstance().player;
     }